reformats Notification class to a react component and refactors notification into its own branch
